code updated

+++ src/app/amc/amc-instrument-renderer.component.ts
@@ -1,9 +1,9 @@
-import {Component} from '@angular/core';
-import {AgRendererComponent} from 'ag-grid-angular';
-import {first} from 'rxjs/operators';
-import {ResultMsg} from "../_models";
-import {AmcinstrumentService} from "../_services/amcinstrument.service";
-import {NotificationService} from "../_services";
+import { Component } from '@angular/core';
+import { AgRendererComponent } from 'ag-grid-angular';
+import { first } from 'rxjs/operators';
+import { ResultMsg } from "../_models";
+import { AmcinstrumentService } from "../_services/amcinstrument.service";
+import { NotificationService } from "../_services";
 
 @Component({
   template: `
@@ -27,13 +27,17 @@ export class AmcInstrumentRendererComponent implements AgRendererComponent {
   }
 
   delete(id) {
-    this.AmcInstrumentService.delete(id)
-      .pipe(first())
-      .subscribe({
-        next: (data: ResultMsg) => {
-          this.notificationService.filter("itemadded");
-        }
-      });
+    console.log(this.params.deleteaccess);
+    
+    if (this.params.deleteaccess) {
+      this.AmcInstrumentService.delete(id)
+        .pipe(first())
+        .subscribe({
+          next: (data: ResultMsg) => {
+            this.notificationService.filter("itemadded");
+          }
+        });
+    }
   }
 
   refresh(): boolean {
